edit 메서드와 마찬가지로 조건문 개선 및 RESTful api 설계원칙에 맞게 메서드와 그에따른 annotation 수정.

+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -60,18 +60,15 @@ public class ArticleController {
         return "qna/show";
     }
 
-    @RequestMapping(value = "/articles/delete/{articleId}", method = {RequestMethod.GET, RequestMethod.DELETE})
+    @DeleteMapping("/articles/delete/{articleId}")
     public String deleteArticle(@PathVariable final Long articleId, HttpServletRequest httpRequest, Model model) {
         HttpSession session = httpRequest.getSession(false);
         ArticleDto article = articleService.findById(articleId);
         model.addAttribute("article", article);
-        if(session == null) return "qna/failed";
-        else if (session.getAttribute("userId").equals(article.getWriter())) {
+        if (session != null && session.getAttribute("userId").equals(article.getWriter())) {
             articleService.deleteArticle(articleId);
             return "redirect:/";
         }
-
         return "qna/failed";
     }
-
 }
